requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# Database / Storage
qdrant-client>=1.10.0
redis>=5.0.0
supabase>=2.5.0

# AI / NLP
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
openai>=1.30.0

# Rate Limiting
slowapi>=0.1.9

# Observability
sentry-sdk[fastapi]>=2.5.0
structlog>=24.2.0

# Billing
stripe>=9.12.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# File / Storage
boto3>=1.34.0
httpx>=0.27.0

# Misc utilities
tenacity>=8.3.0
